tests/requirements.txt
deepeval>=2.0.0
openai>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-html>=4.0.0
